+++ app/src/main/java/kr/com/duri/user/application/service/QuotationReqServiceImpl.java
@@ -1,15 +1,7 @@
 package kr.com.duri.user.application.service;
 
-import java.util.List;
-
-import kr.com.duri.groomer.domain.entity.Quotation;
-import kr.com.duri.groomer.exception.QuotationNotFoundException;
-import kr.com.duri.groomer.repository.QuotationRepository;
 import kr.com.duri.user.domain.entity.QuotationReq;
-import kr.com.duri.user.domain.entity.Request;
-import kr.com.duri.user.exception.RequestNotFoundException;
 import kr.com.duri.user.repository.QuotationReqRepository;
-import kr.com.duri.user.repository.RequestRepository;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
@@ -18,40 +10,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class QuotationReqServiceImpl implements QuotationReqService {
 
-    private final RequestRepository requestRepository;
-    private final QuotationRepository quotationRepository;
     private final QuotationReqRepository quotationReqRepository;
-
-    @Override
-    public Request getRequestById(Long requestId) {
-        return requestRepository
-                .findById(requestId)
-                .orElseThrow(() -> new RequestNotFoundException("견적 요청을 찾을 수 없습니다."));
-    }
-
-    @Override
-    public List<Request> getNewRequestsByShopId(Long shopId) {
-        return requestRepository.findNewRequestsByShopId(shopId);
-    }
-
-    @Override
-    public void updateRequestStatusToApproved(Long requestId) {
-        Request existingRequest = getRequestById(requestId);
-        Request updatedRequest = existingRequest.withApprovedStatus();
-        requestRepository.save(updatedRequest);
-    }
-
-    @Override
-    public List<Request> getApprovedRequestsByShopId(Long shopId) {
-        return requestRepository.findApprovedRequestsByShopId(shopId);
-    }
-
-    @Override
-    public Quotation getQuotationByRequestId(Long requestId) {
-        return quotationRepository
-                .findByRequestId(requestId)
-                .orElseThrow(() -> new QuotationNotFoundException("해당 견적을 찾을 수 없습니다."));
-    }
 
     @Override
     public QuotationReq saveQuotationRequest(QuotationReq quotationReq) {
