+++ app/src/main/java/kr/com/duri/user/application/service/RequestService.java
@@ -6,4 +6,12 @@ import kr.com.duri.user.domain.entity.Request;
 
 public interface RequestService {
     void saveRequests(List<Request> requests);
+
+    Request getRequestById(Long requestId);
+
+    List<Request> getNewRequestsByShopId(Long shopId);
+
+    void updateRequestStatusToApproved(Long requestId);
+
+    List<Request> getApprovedRequestsByShopId(Long shopId);
 }
